Made script running under Windows

+++ Plus/scripts/06_configure_fpga.source.bash
@@ -3,16 +3,19 @@
 
 is_command_available_or_error quartus_pgm " from Intel FPGA Quartus Prime package"
 
-rules_dir=/etc/udev/rules.d
-rules_file="$scripts_dir/90-intel-fpga.rules"
-
-if ! grep -q USB-Blaster $rules_dir/*
+if [ "$OSTYPE" = "linux-gnu" ]
 then
-    error "No rules for USB Blaster detected in $rules_dir."  \
-          "Please put it there and reboot: sudo cp $rules_file $rules_dir"
-fi
+    rules_dir=/etc/udev/rules.d
+    rules_file="$scripts_dir/90-intel-fpga.rules"
 
-killall jtagd 2>/dev/null || true
+    if ! grep -q USB-Blaster $rules_dir/*
+    then
+        error "No rules for USB Blaster detected in $rules_dir."  \
+              "Please put it there and reboot: sudo cp $rules_file $rules_dir"
+    fi
+
+    killall jtagd 2>/dev/null || true
+fi
 
 quartus_pgm -l &> cable_list
